Correct includes value in library.properties
The includes field in library.properties is used to specify which #include directives should be added to the sketch via Sketch > Include Library > ESPHelper. The previous value of this field empty caused that action to add the following line to the sketch:

#include <ESPHelper>

but the correct filename is ESPHelper.h, not ESPHelper so that would cause a compilation error.